Split out examples to own package
