+++ package.json
@@ -3,7 +3,8 @@
     "private": true,
     "workspaces": [
         "packages/matter.js",
-        "packages/matter-node.js"
+        "packages/matter-node.js",
+        "packages/matter-node.js-examples"
     ],
     "scripts": {
         "prepare": "npm run build-clean",
@@ -15,10 +16,10 @@
         "lint-fix": "eslint --fix **/*.ts",
         "format": "tsfmt --useTsconfig tsconfig.tsfmt.lint.json -r",
         "format-verify": "tsfmt --useTsconfig tsconfig.tsfmt.lint.json --verify",
-        "matter-device": "ts-node --project packages/matter-node.js/tsconfig.dist.json packages/matter-node.js/src/examples/DeviceNode.ts",
-        "matter-bridge": "ts-node --project packages/matter-node.js/tsconfig.dist.json packages/matter-node.js/src/examples/BridgedDevicesNode.ts",
-        "matter-composeddevice": "ts-node --project packages/matter-node.js/tsconfig.dist.json packages/matter-node.js/src/examples/ComposedDeviceNode.ts",
-        "matter-controller": "ts-node --project packages/matter-node.js/tsconfig.dist.json packages/matter-node.js/src/examples/ControllerNode.ts"
+        "matter-device": "ts-node --project packages/matter-node.js-examples/tsconfig.dist.json packages/matter-node.js-examples/src/examples/DeviceNode.ts",
+        "matter-bridge": "ts-node --project packages/matter-node.js-examples/tsconfig.dist.json packages/matter-node.js-examples/src/examples/BridgedDevicesNode.ts",
+        "matter-composeddevice": "ts-node --project packages/matter-node.js-examples/tsconfig.dist.json packages/matter-node.js-examples/src/examples/ComposedDeviceNode.ts",
+        "matter-controller": "ts-node --project packages/matter-node.js-examples/tsconfig.dist.json packages/matter-node.js-examples/src/examples/ControllerNode.ts"
     },
     "devDependencies": {
         "@typescript-eslint/eslint-plugin": "^5.59.6",
