+++ packages/matter-node.js-examples/.npmignore
@@ -1,0 +1,2 @@
+/tsconfig.json
+/src
