+++ packages/matter-node.js-examples/package.json
@@ -1,0 +1,66 @@
+{
+    "name": "@project-chip/matter-node.js-examples",
+    "version": "0.4.0",
+    "description": "CLI/Reference implementation scripts for Matter protocol for node.js",
+    "keywords": [
+        "iot",
+        "home automation",
+        "matter",
+        "smart device"
+    ],
+    "license": "Apache-2.0",
+    "author": "Project CHIP Authors",
+    "contributors": [
+        "Martin Turon <mturon@google.com>",
+        "Marco Fucci di Napoli <mfucci@gmail.com>",
+        "Ingo Fischer <github@fischer-ka.de>"
+    ],
+    "bugs": {
+        "url": "https://github.com/project-chip/matter.js/issues"
+    },
+    "homepage": "https://github.com/project-chip/matter.js",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/project-chip/matter.js.git"
+    },
+    "scripts": {
+        "clean": "rm -rf dist && rm -f *.tsbuildinfo",
+        "build": "tsc -b tsconfig.dist.json",
+        "build-clean": "npm run clean && npm run build",
+        "build-doc": "true",
+        "matter-device": "ts-node --project tsconfig.dist.json src/examples/DeviceNode.ts",
+        "matter-bridge": "ts-node --project tsconfig.dist.json src/examples/BridgedDevicesNode.ts",
+        "matter-composeddevice": "ts-node --project tsconfig.dist.json src/examples/ComposedDeviceNode.ts",
+        "matter-controller": "ts-node --project tsconfig.dist.json src/examples/ControllerNode.ts"
+    },
+    "bin": {
+        "matter-device": "./dist/examples/DeviceNode.js",
+        "matter-bridge": "./dist/examples/BridgedDevicesNode.js",
+        "matter-composeddevice": "./dist/examples/ComposedDeviceNode.js",
+        "matter-controller": "./dist/examples/ControllerNode.js"
+    },
+    "devDependencies": {
+        "@typescript-eslint/eslint-plugin": "^5.59.6",
+        "@typescript-eslint/parser": "^5.59.6",
+        "eslint": "^8.40.0",
+        "eslint-import-resolver-typescript": "^3.5.5",
+        "eslint-plugin-import": "^2.27.5",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5"
+    },
+    "dependencies": {
+        "@project-chip/matter-node.js": "^0.4.0"
+    },
+    "engines": {
+        "_comment": "For Crypto.hkdf support",
+        "node": ">=16.0.0"
+    },
+    "files": [
+        "dist/**/*",
+        "LICENSE",
+        "README.md"
+    ],
+    "publishConfig": {
+        "access": "public"
+    }
+}
